data_collection/requirements.txt
paho_mqtt==1.5.0
pyserial==3.5
Requests==2.31.0
tomli==2.0.1
sparkfun-ublox-gps
spidev